refactor(map): extract player marker logic to separate file
- Move player-related UI components and state from MainWindow.xaml.cs to dedicated MainWindow.Map.Players.cs
- Remove unused fields (_mapReady, _lastHost, _panLastWorld) and legacy shop hover popup code
- Simplify device list XAML by restructuring layout and removing redundant display logic
- Add _shopIconSet field to MainWindow.Map.Shops.cs for shop icon tracking
