+++ RustPlusDesktop/MainWindow.xaml.cs
@@ -2118,17 +2118,6 @@ public partial class MainWindow : Window
         var s = Math.Pow(effZoom, -exp) * baseMult;
         return Math.Clamp(s, ICON_SCALE_MIN, ICON_SCALE_MAX);
     }
-    private Popup? _shopsHoverPopup;
-    private WrapPanel? _shopsHoverWrap;
-    private readonly HashSet<FrameworkElement> _shopIconSet = new(); // alle Icon-Roots
-    private void ApplyMonumentScale(FrameworkElement el)
-    {
-        if (el == null) return;
-        double eff = GetEffectiveZoom();               // dein eff ist perfekt
-        double scale = CalcOverlayScale(eff, MON_SIZE_EXP, MON_BASE_MULT);
-        el.RenderTransformOrigin = new Point(0.5, 0.5);
-        el.RenderTransform = new ScaleTransform(scale, scale);
-    }
 
     private static readonly Dictionary<string, string> sMonIconByKey =
     BuildCanonIconMap(sMonIconByKeyRaw);
@@ -4633,470 +4622,6 @@ private sealed record MarkerRef(System.Windows.Shapes.Ellipse Dot, double U_DIP,
         }
     }
 
-    // avatar size on the map
-    private const double PlayerAvatarSize = 24;
-
-    private sealed class PlayerMarkerTag
-    {
-        public ulong SteamId;
-        public TextBlock NameText = null!;
-        public string? Name { get; set; }
-        public Ellipse? AvatarCircle;
-        public double Radius;
-        public bool IsDeathPin { get; set; }
-        public bool IsDot;
-
-        // Per-Element-Scaling:
-        public double ScaleExp { get; set; } = SHOP_SIZE_EXP; // stärker/schwächer je Typ
-        public double ScaleBaseMult { get; set; } = 1.0;      // Grundgröße
-        public FrameworkElement? ScaleTarget { get; set; }
-        public double ScaleCenterX { get; set; }  // Pivot X
-        public double ScaleCenterY { get; set; }  // Pivot Y
-        public double Rotation { get; set; }      // NEU: für orientierte Marker
-        public TextBlock? TimerText { get; set; }
-        public FrameworkElement? TimerContainer { get; set; }
-    }
-
-
-
-
-
-    // quick lookup from your Team list (no extra web calls here)
-    private ImageSource? GetAvatar(ulong sid)
-        => TeamMembers.FirstOrDefault(t => t.SteamId == sid)?.Avatar;
-
-    private ImageSource? GetAvatarForMap(ulong sid)
-    => _showProfileMarkers ? GetTeamAvatar(sid) : null;
-
-    private ImageSource? GetTeamAvatar(ulong sid)
-    {
-        var vm = TeamMembers.FirstOrDefault(t => t.SteamId == sid);
-        if (vm?.Avatar != null) return vm.Avatar;
-
-        if (_avatarCache.TryGetValue(sid, out var img) && img != null)
-            return img;
-        return null;
-
-    }
-
-    // Small dot + name label (used when Profile markers is OFF)
-    private FrameworkElement BuildPlayerDotMarker(ulong sid, string name, bool online, bool dead)
-    {
-        var brush = dead ? Brushes.IndianRed : (online ? Brushes.LimeGreen : Brushes.LightGray);
-
-        var dot = new Ellipse
-        {
-            Width = 10,
-            Height = 10,
-            Fill = brush,
-            Stroke = Brushes.Black,
-            StrokeThickness = 2,
-            Margin = new Thickness(0, 0, 4, 0),
-
-        };
-
-        var tb = new TextBlock
-        {
-            Text = name,
-            Foreground = brush,
-            FontSize = 12,
-            Margin = new Thickness(6, -2, 0, 0)
-        };
-
-        var sp = new StackPanel { Orientation = Orientation.Horizontal };
-        sp.Children.Add(dot);
-        sp.Children.Add(tb);
-        ToolTipService.SetToolTip(sp, name);
-
-        // keep Radius=5 so centering uses the dot’s radius, not the label width
-        sp.Tag = new PlayerMarkerTag
-        {
-            SteamId = sid,
-            Name = name,
-            NameText = tb,
-            AvatarCircle = null,
-            Radius = 5,
-            IsDeathPin = false,
-            IsDot = true, // <— new flag to distinguish dot vs avatar
-            ScaleExp = 1.05,
-            ScaleBaseMult = 1.0,
-            ScaleTarget = sp,          // << gesamten Inhalt skalieren (Dot + Name)
-            ScaleCenterX = 5.0,        // << Dot.Width / 2.0 (10/2)
-            ScaleCenterY = 5.0         // << Dot.Height / 2.0
-        };
-
-        return sp;
-    }
-
-    // build a player marker (avatar circle if available, dot otherwise)
-    private FrameworkElement BuildPlayerMarker(ulong sid, string name, bool online, bool dead)
-    {
-        var brush = dead ? Brushes.IndianRed : (online ? Brushes.LimeGreen : Brushes.Gray);
-        var avatar = GetAvatar(sid);
-
-        if (avatar == null)
-        {
-            var dot = new Ellipse
-            {
-                Width = 10,
-                Height = 10,
-                Fill = brush,
-                Stroke = Brushes.Black,
-                StrokeThickness = 2,
-                Margin = new Thickness(0, 0, 4, 0)
-            };
-            var tb = new TextBlock { Text = name, Foreground = brush, FontSize = 12, Margin = new Thickness(6, -2, 0, 0) };
-            var sp = new StackPanel { Orientation = Orientation.Horizontal };
-            sp.Children.Add(dot);
-            sp.Children.Add(tb);
-            sp.Tag = new PlayerMarkerTag
-            {
-                SteamId = sid,
-                NameText = tb,
-                AvatarCircle = null,
-                Radius = 5,
-                IsDot = true,
-                ScaleExp = 1.05,
-                ScaleBaseMult = 1.0,
-                ScaleTarget = sp,
-                ScaleCenterX = 5.0,
-                ScaleCenterY = 5.0
-            };
-            Panel.SetZIndex(sp, 905);
-            ApplyCurrentOverlayScale(sp);
-            return sp;
-        }
-        else
-        {
-            var tb = new TextBlock { Text = name, Foreground = brush, FontSize = 12, Margin = new Thickness(6, -2, 0, 0) };
-            var circle = new Ellipse
-            {
-                Width = PlayerAvatarSize,
-                Height = PlayerAvatarSize,
-                Stroke = Brushes.Black,
-                StrokeThickness = 2,
-                Fill = new ImageBrush(avatar) { Stretch = Stretch.UniformToFill }
-            };
-
-            var host = new Grid();
-            host.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto });
-            host.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto });
-
-            var avatarHost = new Grid { Width = PlayerAvatarSize, Height = PlayerAvatarSize, Margin = new Thickness(0, 0, 4, 0) };
-            avatarHost.Children.Add(circle);
-
-            host.Children.Add(avatarHost);
-            Grid.SetColumn(avatarHost, 0);
-            host.Children.Add(tb);
-            Grid.SetColumn(tb, 1);
-
-            host.Tag = new PlayerMarkerTag
-            {
-                SteamId = sid,
-                NameText = tb,
-                AvatarCircle = circle,
-                Radius = PlayerAvatarSize * 0.5,
-                ScaleExp = 0.85,
-                ScaleBaseMult = 1.0,
-                ScaleTarget = host,              // << beide (Avatar + Name) skalieren
-                ScaleCenterX = PlayerAvatarSize * 0.5, // << Zentrum des Avatars (links)
-                ScaleCenterY = PlayerAvatarSize * 0.5,
-
-            };
-            Panel.SetZIndex(host, 905);
-            ToolTipService.SetToolTip(host, name);
-            ApplyCurrentOverlayScale(host);
-
-            return host;
-        }
-    }
-
-    private readonly HashSet<ulong> _avatarLoading = new();
-    private readonly Dictionary<ulong, DateTime> _avatarNextTry = new();
-    private static readonly TimeSpan AvatarRetryInterval = TimeSpan.FromSeconds(30);
-
-    private bool CanTryAvatar(ulong sid)
-    {
-        if (_avatarLoading.Contains(sid)) return false;
-        return !_avatarNextTry.TryGetValue(sid, out var next) || DateTime.UtcNow >= next;
-    }
-
-    private FrameworkElement MakePlayerDot(string tooltip, bool online)
-    {
-        var dot = new Ellipse
-        {
-            Width = 10,
-            Height = 10,
-            Fill = online ? Brushes.LimeGreen : Brushes.LightGray,
-            Stroke = Brushes.Black,
-            StrokeThickness = 2,
-            Margin = new Thickness(0, 0, 4, 0),
-
-        };
-        ToolTipService.SetToolTip(dot, tooltip);
-        Panel.SetZIndex(dot, 905);
-        return dot;
-    }
-
-    // update existing marker (swap to avatar when it becomes available, recolor & rename)
-    private void UpdatePlayerMarker(ref FrameworkElement el, uint key, ulong sid, string name, bool online, bool dead)
-    {
-        if (sid == 0) return; // niemals Events o.ä. „umwandeln“
-        // Checkbox respektieren:
-        if (!_showProfileMarkers)
-        {
-            var brush = dead ? Brushes.IndianRed : (online ? Brushes.LimeGreen : Brushes.LightGray);
-
-            // Wenn noch kein "Dot+Name"-Marker existiert (oder ein Avatar aktiv war): neu bauen
-            if (el.Tag is not PlayerMarkerTag t || !t.IsDot)
-            {
-                var newEl = BuildPlayerDotMarker(sid, name, online, dead);
-                int idx = Overlay.Children.IndexOf(el);
-                if (idx >= 0) { Overlay.Children.RemoveAt(idx); Overlay.Children.Insert(idx, newEl); }
-                else Overlay.Children.Add(newEl);
-                _dynEls[key] = newEl; el = newEl;
-                Panel.SetZIndex(newEl, 905);
-            }
-            else
-            {
-                // Bereits ein Dot-StackPanel vorhanden -> nur Name/Brush aktualisieren
-                t.NameText.Text = name;
-                t.NameText.Foreground = brush;
-
-                // Dot (erste Ellipse im StackPanel) umfärben
-                if (t.NameText.Parent is Panel sp)
-                {
-                    var dot = sp.Children.OfType<Ellipse>().FirstOrDefault();
-                    Panel.SetZIndex(dot, 905);
-                    if (dot != null) dot.Fill = brush;
-                }
-            }
-
-            // Tooltip konsistent halten (optional)
-            ToolTipService.SetToolTip(el, name);
-            return; // <— wichtig
-        }
-
-        // (Rest: deine bisherige Logik für Avatar-/Name-Update)
-        var brush2 = dead ? Brushes.IndianRed : (online ? Brushes.LimeGreen : Brushes.LightGray);
-        var avatar = GetAvatarForMap(sid);
-
-        if (el.Tag is PlayerMarkerTag tag)
-        {
-            if (tag.NameText != null) tag.NameText.Text = name;
-            if (tag.NameText != null) tag.NameText.Foreground = brush2;
-
-            if (avatar != null && tag.AvatarCircle == null ||
-                avatar == null && tag.AvatarCircle != null)
-            {
-                var newEl = BuildPlayerMarker(sid, name, online, dead);
-                int idx = Overlay.Children.IndexOf(el);
-                if (idx >= 0) { Overlay.Children.RemoveAt(idx); Overlay.Children.Insert(idx, newEl); }
-                else Overlay.Children.Add(newEl);
-                _dynEls[key] = newEl; el = newEl;
-            }
-            else if (tag.AvatarCircle != null && avatar != null)
-            {
-                tag.AvatarCircle.Fill = new ImageBrush(avatar) { Stretch = Stretch.UniformToFill };
-            }
-        }
-        else
-        {
-            var newEl = BuildPlayerMarker(sid, name, online, dead);
-            int idx = Overlay.Children.IndexOf(el);
-            if (idx >= 0) { Overlay.Children.RemoveAt(idx); Overlay.Children.Insert(idx, newEl); }
-            else Overlay.Children.Add(newEl);
-            _dynEls[key] = newEl; el = newEl;
-        }
-    }
-
-    private void ChkProfileMarkers_Toggled(object? sender, RoutedEventArgs e)
-    {
-        _showProfileMarkers = ChkProfileMarkers.IsChecked == true;
-
-        // alle existierenden Player-Marker umschalten
-        foreach (var kv in _dynEls.ToList())
-        {
-            if (kv.Value is FrameworkElement el && el.Tag is PlayerMarkerTag tag)
-            {
-                if (tag.SteamId == 0 || tag.IsDeathPin) continue; // keine Events/DeathPins anfassen
-                var sid = tag.SteamId;
-                var name = TeamMembers.FirstOrDefault(t => t.SteamId == sid)?.Name ?? "player";
-                //var st = _lastPresence.TryGetValue(sid, out var p) ? p : (false, false);
-                if (_lastPresence.TryGetValue(sid, out var p))
-                {
-                    var online = p.Item1;
-                    var dead = p.Item2;
-                    UpdatePlayerMarker(ref el, kv.Key, sid, name, online, dead);
-                }
-                else
-                {
-                    UpdatePlayerMarker(ref el, kv.Key, sid, name, online: false, dead: false);
-                }
-            }
-        }
-    }
-
-    private void ChkDeathMarkers_Toggled(object? sender, RoutedEventArgs e)
-    {
-        _showDeathMarkers = ChkDeathMarkers.IsChecked == true;
-        if (!_showDeathMarkers) ClearAllDeathPins(); // „Deaktivieren cleart“
-    }
-
-    private const double PinW = 40;   // Gesamtbreite
-    private const double PinH = 56;   // Gesamthöhe (inkl. Spitze)
-    private const double Circle = 24; // Durchmesser der Kreisfläche im Kopf
-    private const double CircleTop = 6; // Abstand von oben bis Kreis-OBERKANTE
-
-
-    private void RefreshAllOverlayScales()
-    {
-        // Player & sonstige dynamische Elemente
-        foreach (var fe in _dynEls.Values)
-            ApplyCurrentOverlayScale(fe);
-
-        // DeathPins
-        foreach (var fe in _deathPins.Values)
-            ApplyCurrentOverlayScale(fe);
-
-        // Shops (falls du sie weiter separat halten willst)
-        RefreshShopIconScales();
-    }
-    private FrameworkElement BuildDeathPin(ulong sid, string name, ImageSource? avatarFromCaller = null)
-    {
-        var avatar = GetTeamAvatar(sid);
-
-        var root = new Grid
-        {
-            Width = PinW,
-            Height = PinH,
-
-            Tag = new PlayerMarkerTag
-            {
-                SteamId = sid,
-                Name = name,
-                IsDeathPin = true,
-
-                // wichtig fürs Skalieren:
-                ScaleExp = 0.8,
-                ScaleBaseMult = 0.9,
-                ScaleTarget = null,                 // null => wir skalieren root selbst
-                ScaleCenterX = PinW * 0.5,          // Pivot X = Mitte unten (Spitze)
-                ScaleCenterY = PinH                 // Pivot Y = Unterkante (Spitze)
-            }
-        };
-
-        // rote Tropfenform (Path ist auf Containergröße gestretcht)
-        var pinPath = Geometry.Parse(
-            "M20,0 C31,0 40,9 40,20 C40,33 20,56 20,56 C20,56 0,33 0,20 C0,9 9,0 20,0 Z"
-        );
-
-        var fill = TryFindResource("DeathPinFill") as Brush ?? Brushes.IndianRed;
-
-        root.Children.Add(new System.Windows.Shapes.Path
-        {
-            Data = pinPath,
-            Fill = fill,
-            Stroke = Brushes.Black,
-            StrokeThickness = 2,
-            Stretch = Stretch.Fill,
-            Width = PinW,
-            Height = PinH
-        });
-
-        // schwarzer Ring um die Kreisfläche
-        root.Children.Add(new Ellipse
-        {
-            Width = Circle + 6,
-            Height = Circle + 6,
-            Stroke = Brushes.Black,
-            StrokeThickness = 2,
-            Fill = Brushes.Transparent,
-            HorizontalAlignment = HorizontalAlignment.Left,
-            VerticalAlignment = VerticalAlignment.Top,
-            Margin = new Thickness((PinW - (Circle + 6)) / 2.0, CircleTop - 3, 0, 0)
-        });
-
-        // Avatar (kreisförmig geclippt) – HORIZONTAL zentriert, TOP = CircleTop
-        FrameworkElement avatarEl;
-        if (avatar != null)
-        {
-            var holder = new Grid { Width = Circle, Height = Circle };
-            holder.Clip = new EllipseGeometry(new Point(Circle / 2.0, Circle / 2.0), Circle / 2.0, Circle / 2.0);
-            holder.Children.Add(new Image { Source = avatar, Stretch = Stretch.UniformToFill });
-            avatarEl = holder;
-        }
-        else
-        {
-            avatarEl = new Ellipse { Width = Circle, Height = Circle, Fill = Brushes.Gray };
-        }
-
-        avatarEl.HorizontalAlignment = HorizontalAlignment.Left;
-        avatarEl.VerticalAlignment = VerticalAlignment.Top;
-        avatarEl.Margin = new Thickness((PinW - Circle) / 2.0, CircleTop, 0, 0);
-        root.Children.Add(avatarEl);
-
-        // Tooltip
-        ToolTipService.SetToolTip(root, $"{name} (death)");
-        ApplyCurrentOverlayScale(root);
-        return root;
-    }
-
-    private void PlaceDeathPin(TeamMemberVM vm)
-    {
-        if (!_showDeathMarkers) return;
-        if (!(vm.X.HasValue && vm.Y.HasValue)) return;
-
-        var px = WorldToImagePx(vm.X.Value, vm.Y.Value);
-        var el = BuildDeathPin(vm.SteamId, vm.Name, GetTeamAvatar(vm.SteamId));
-
-        // ersetzen, falls derselbe Spieler erneut stirbt
-        if (_deathPins.TryGetValue(vm.SteamId, out var old))
-        {
-            Overlay.Children.Remove(old);
-            _deathPins.Remove(vm.SteamId);
-        }
-
-        Overlay.Children.Add(el);
-        Panel.SetZIndex(el, 805); // über Events, unter Namen
-        ApplyCurrentOverlayScale(el);
-
-        // Pinspitze auf Position
-        var cx = px.X - PinW / 2.0;
-        var cy = px.Y - (CircleTop + Circle / 2.0);
-        Canvas.SetLeft(el, cx);
-        Canvas.SetTop(el, cy);
-
-        _deathPins[vm.SteamId] = el;
-    }
-
-    private void PlaceOrMoveDeathPin(ulong sid, double worldX, double worldY, string name)
-    {
-        var px = WorldToImagePx(worldX, worldY);
-        var el = _deathPins.TryGetValue(sid, out var exist) ? exist : null;
-
-        if (el == null)
-        {
-            el = BuildDeathPin(sid, name);
-            _deathPins[sid] = el;
-            Overlay.Children.Add(el);
-            Panel.SetZIndex(el, 805);
-            ApplyCurrentOverlayScale(el);
-        }
-
-        // Spitze auf Position
-        var cx = px.X - (PinW / 2.0);
-        var cy = px.Y - PinH;
-        Canvas.SetLeft(el, cx);
-        Canvas.SetTop(el, cy);
-    }
-
-    private void ClearAllDeathPins()
-    {
-        foreach (var kv in _deathPins) Overlay.Children.Remove(kv.Value);
-        _deathPins.Clear();
-    }
-
     private async Task SendTeamChatSafeAsync(string text)
     {
         // Thread-safe: called from both UI thread (DynEvents) and background threads (MonumentWatcher)
@@ -5134,117 +4659,6 @@ private sealed record MarkerRef(System.Windows.Shapes.Ellipse Dot, double U_DIP,
         7 => "Building Blocked",
         _ => "Event"
     };
-
-    private bool _didInitialOverlayRefresh;
-
-    private async void EnsureInitialOverlayRefresh()
-    {
-        if (_didInitialOverlayRefresh || Overlay == null) return;
-        _didInitialOverlayRefresh = true;
-
-        // Einen Tick warten, bis Measure/Arrange/Render einmal durch sind
-        await Dispatcher.Yield(System.Windows.Threading.DispatcherPriority.Loaded);
-        // alternativ: DispatcherPriority.Render
-
-        RefreshAllOverlayScales();          // skaliert Spieler/Death/Shops
-        RefreshMonumentOverlayPositions();  // falls du Monumente separat setzt
-    }
-
-    // wie stark du die Größe kompensierst:
-    // 0.0 = gar nicht (verhält sich wie Map), 1.0 = am Bildschirm immer gleich groß,
-    // 0.7–0.9 = "etwas" größer raus / etwas kleiner rein
-    private const double SHOP_SIZE_EXP = 0.8;
-
-    // Gesamtzoom (Viewbox * MapTransform)
-    private double GetEffectiveZoom()
-    {
-        var (s, _, _) = GetViewboxScaleAndOffset(); // hast du schon
-        var m = MapTransform.Matrix;
-        double eff = Math.Abs(s * m.M11);
-        return eff <= 1e-6 ? 1e-6 : eff;
-    }
-
-    // auf ein einzelnes Shop-Element anwenden
-    private void ApplyCurrentOverlayScale(FrameworkElement el)
-    {
-        if (el == null) return;
-
-        double eff = GetEffectiveZoom();
-        double exp = SHOP_SIZE_EXP, baseMult = SHOP_BASE_MULT;
-
-        FrameworkElement target = el;
-        double centerX = -1.0, centerY = -1.0;
-
-        if (el.Tag is PlayerMarkerTag pt)
-        {
-            if (pt.ScaleExp > 0) exp = pt.ScaleExp;
-            if (pt.ScaleBaseMult > 0) baseMult = pt.ScaleBaseMult;
-
-            // für DeathPins wollen wir root skalieren:
-            if (pt.IsDeathPin)
-            {
-                target = el;
-                centerX = pt.ScaleCenterX;   // = PinW/2
-                centerY = pt.ScaleCenterY;   // = PinH
-            }
-            else if (pt.ScaleTarget != null)
-            {
-                // (Spieler: Avatar/Dot+Name etc.)
-                target = pt.ScaleTarget;
-                centerX = pt.ScaleCenterX;
-                centerY = pt.ScaleCenterY;
-
-                if (!ReferenceEquals(target, el))
-                    el.RenderTransform = Transform.Identity; // Root nicht zusätzlich skalieren
-            }
-        }
-
-        double scale = CalcOverlayScale(eff, exp, baseMult);
-        double rotation = (el.Tag is PlayerMarkerTag ptRot) ? ptRot.Rotation : 0;
-
-        // Pivot-Logik: Falls wir explizite Center-Koordinaten haben, nutzen wir diese
-        if (centerX >= 0 && centerY >= 0)
-        {
-            double w = target.ActualWidth > 0 ? target.ActualWidth : target.Width;
-            double h = target.ActualHeight > 0 ? target.ActualHeight : target.Height;
-
-            if (w > 0 && h > 0)
-            {
-                target.RenderTransformOrigin = new Point(centerX / w, centerY / h);
-            }
-            else
-            {
-                target.RenderTransformOrigin = new Point(0.5, 0.5);
-            }
-        }
-        else
-        {
-            target.RenderTransformOrigin = new Point(0.5, 0.5); 
-        }
-        
-        var group = target.RenderTransform as TransformGroup;
-        if (group == null || group.Children.Count < 2 || !(group.Children[0] is ScaleTransform) || !(group.Children[1] is RotateTransform))
-        {
-            group = new TransformGroup();
-            group.Children.Add(new ScaleTransform(scale, scale));
-            group.Children.Add(new RotateTransform(rotation));
-            target.RenderTransform = group;
-        }
-        else
-        {
-            var st = (ScaleTransform)group.Children[0];
-            st.ScaleX = scale;
-            st.ScaleY = scale;
-
-            // Only update rotation if not currently animating it (handled in AnimateMarkerRotation)
-            var rt = (RotateTransform)group.Children[1];
-            var source = DependencyPropertyHelper.GetValueSource(rt, RotateTransform.AngleProperty);
-            if (!source.IsAnimated)
-            {
-                rt.Angle = rotation;
-            }
-        }
-    }
 
     private class TwoStepFlip
     {
